Add: Delete and Put API requests

+++ GaelO-Flow/src/users/users.controller.ts
@@ -20,7 +20,6 @@ export class UsersController {
 
   @Get()
   async getUsers(): Promise<User[]> {
-    console.log(await this.UserService.findAll());
     return await this.UserService.findAll();
   }
 
@@ -29,22 +28,50 @@ export class UsersController {
     return await this.UserService.findOne(id);
   }
 
-  @Put('/:id')
+  @Put('/:id') //TODO
   async update(
     @Param('id') id: number,
     @Body() userDto: UserDto,
   ): Promise<void> {
-    // return await this.UserService.update(id, body);
+    const user = await this.UserService.findOne(id);
+    const regexEmail = /^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$/g;
+    const regexPassword =
+      /^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$ %^&*-]).{12,}$/g;
+
+    if (!user) throw new HttpException('User not found', 404);
+
+    if (userDto.firstname) user.firstname = userDto.firstname;
+    if (userDto.lastname) user.lastname = userDto.lastname;
+    if (userDto.username) user.username = userDto.username;
+    if (userDto.password) {
+      if (regexPassword.test(userDto.password) === false)
+        throw new HttpException('Password is not valid', 400);
+      const salt = await bcrypt.genSalt();
+      const hash = await bcrypt.hash(userDto.password, salt);
+      user.password = hash;
+      user.salt = salt;
+    }
+    if (userDto.email) {
+      if (regexEmail.test(userDto.email) === false)
+        throw new HttpException('Email is not valid', 400);
+      user.email = userDto.email;
+    }
+    if (userDto.super_admin) user.super_admin = userDto.super_admin;
+    if (userDto.role_name) user.role_name = userDto.role_name;
+    if (userDto.is_active) user.is_active = userDto.is_active;
+
+    await this.UserService.update(id, user);
   }
 
   @Delete('/:id')
   async delete(@Param('id') id: number): Promise<void> {
+    const user = await this.UserService.findOne(id);
+    if (!user) throw new HttpException('User not found', 404);
     return await this.UserService.remove(id);
   }
 
   @Post()
   async createUser(@Body() userDto: UserDto): Promise<number> {
-    const role = new Role();
     const user = new User();
     const regexEmail = /^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$/g;
     const regexPassword =
@@ -53,8 +80,6 @@ export class UsersController {
     const hash = await bcrypt.hash(userDto.password, salt);
 
     // const isMatch = await bcrypt.compare(userDto.password, hash);
-
-    role.name = userDto.role_name;
 
     if (regexEmail.test(userDto.email) === false)
       throw new HttpException('Email is not valid', 400);
@@ -72,10 +97,16 @@ export class UsersController {
     user.role_name = userDto.role_name;
     user.salt = salt;
 
-    const existingUser = await this.UserService.findByUsernameOrEmail(userDto.username, userDto.email);
+    const existingUser = await this.UserService.findByUsernameOrEmail(
+      userDto.username,
+      userDto.email,
+    );
 
     if (existingUser) {
-      throw new HttpException('User with this username or email already exists', 409)
+      throw new HttpException(
+        'User with this username or email already exists',
+        409,
+      );
     }
 
     return await this.UserService.create(user);
